Unify the brush-mask Metal pipeline construction
The brush-mask render pipeline was constructed twice: once inline in
`_EditingCanvasMTKView` for live painting, and once in
`BrushMaskMetalRasterizer` for the off-screen path that
`BrushMaskRasterizerParityTests` exercises. The two copies covered the
shader library load, the pipeline descriptor (pixel format, blend state),
the `BrushStampUniforms` layout, and the per-stamp encoding, and were held
equal only by doc comments asserting they were identical.

That made the parity test weaker than it looks: it proves the *rasterizer's*
copy matches the parametric Core Image kernel, not the pipeline the live
view actually drives. A drift in the live view's blend op, pixel format, or
uniform layout would have left the test green.

These two constructions must evolve together — they target the same shader
functions, and the uniform layout is ABI against the MSL `BrushStampUniforms`
struct in `BrushMaskRenderShader.metal`. Extract them into one internal
`BrushMaskPipeline` factory (library, pipeline state, `StampUniforms`, and
`encodeStamp`) that both paths call, so the "identical by construction"
contract is structural rather than comment-enforced, and the parity test now
covers the live path's construction by construction.

No behavior change: the two copies were already effect-identical (same
`.rgba8Unorm` attachment, same `.max` rgb/alpha blend ops with all `.one`
factors, same field-for-field uniform layout, same
`setVertexBytes`/`setFragmentBytes` at buffer index 0 with
`MemoryLayout<...>.stride`, same 4-vertex triangle-strip draw). Only the
prose of their comments differed.
